XW-849 rename all files in common/tests, common/modules and common/baseline
