+++ front/auth/app/tracking.js
@@ -1,5 +1,5 @@
-import AuthTracker from './common/AuthTracker';
-import UniversalAnalytics from 'common/modules/Trackers/UniversalAnalytics';
+import AuthTracker from './common/auth-tracker';
+import UniversalAnalytics from 'common/modules/trackers/universal-analytics';
 import {trackActions} from 'common/utils/track';
 import {getQueryParam} from 'common/utils/queryString';
