+++ front/common/baseline/load-dom-resource.js
