+++ front/common/modules/instant-globals.js
