+++ front/common/modules/thumbnailer.js
