+++ front/common/modules/trackers/base-tracker.js
