+++ front/common/modules/video-players/base.js
