+++ front/main/app/instance-initializers/setup-tracking.js
@@ -1,7 +1,7 @@
-import Ads from 'common/modules/Ads';
-import UniversalAnalytics from 'common/modules/Trackers/UniversalAnalytics';
+import Ads from 'common/modules/ads';
+import UniversalAnalytics from 'common/modules/trackers/universal-analytics';
 import {integrateOptimizelyWithUA} from 'common/utils/variantTesting';
-import {integrateAbTestWithUA} from 'common/modules/AbTest';
+import {integrateAbTestWithUA} from 'common/modules/abtest';
 
 /**
  * @returns {void}
